Simplify ping interface
Reduce the number of ping functions. There are 2 types now: a version
(ping/1, ping/2) that opens/closes the socket for the user and ping/3
that takes a socket as its first option.

All other options can be specified as a proplist to ping/2 or ping/3.
Retain ping/1 since it's convenient for one-off/infrequent usage. Even
the socket could be specified as an option but it feels a bit unnatural.