Optionally do not fork helper to get raw socket
gen_icmp uses procket to get the raw socket for ping'ing. For safety,
procket calls a setuid helper program to create the socket. If it's ok
to run beam with escalated privs, gen_icmp has the option of requesting
the socket within Erlang.

Thanks to @uwe_ on twitter for the suggestion!