Add option to open an IPv6 socket
Add an inet6 option similar to gen_tcp/gen_udp. Note this is just a
shortcut; it was already possible to open an IPv6 socket by using open/2:

    gen_icmp:open([{family, inet6}], []).

open/1 was introduced to simplify:

    gen_icmp:open([inet6]).

Arity is 1, deviating from gen_tcp/gen_udp, because there is no value
needed for the port and type/code can be set on a per packet basis.