Suppress informational messages from gen_server
At some point, the erlang open_port mechanism for passing in file
descriptors seems to have changed. When the fd is closed, an exit
message is generated.

It also seems that a port is opened for each fd:

    erlang:open_port({fd,1,1}, [binary,stream])

One of the ports exits immediately.

Modify the info callback to ignore exit messages.