Fix compile failures caused by changes in pkt
Removing the PF_INET6 macro from pkt broke building tracert.
Conditionally include a macro that figures out the value of PF_INET6 at
runtime. A side effect of this change is that tracrt should now work on
platforms other than Linux.

The function to determine the value of PF_INET6 (tracert:family/1) should
be centralized in procket in the future. procket has a similar function
but it is not exported.

Pin the dependencies to tags on procket and pkt as suggested by
@divolgin in:

https://github.com/msantos/pkt/commit/41a99073655285d21dacaae89196bed396e86cff

Reported by @tnt-dev in:

https://github.com/msantos/gen_icmp/issues/7

Thanks divolgin/tnt-dev!