Return name resolution errors
ping/1,2 would fail with a badmatch exception if the host failed to
resolve. A temporary name resolution error in one of a list of hosts
would cause the entire operation to abort:

https://github.com/msantos/gen_icmp/issues/10

Since tracert:host/1 does not return an ok/error tuple, continue to crash
on name resolution failures.

Thanks @tchoutri!