Remove the dedup option
Do not minimize the list of hosts passed in to gen_icmp:ping/1,2.

The 'dedup' option made sense when the host list was constant. Hosts
could be pulled out of the results using the proplists functions.

When name resolution is used, the interaction of the host list with the
dedup option is ambiguous. For example:

    % google.com = 173.194.46.34
    % Should dedup be based on hostname or resolved IP address?
    gen_icmp:ping(["google.com", "173.194.46.34", {173,194,46,34}],
            [dedup])

Name resolution failures make the situation even worse:

    % e.g., returns: 1.1.1.1, nxdomain, 1.1.1.2
    % which should be used: one of the successful name resolutions,
    % the resolution failure or all three?
    gen_icmp:ping(["google.com", "google.com", "google.com"], [dedup])

When resolving hostnames, do not sort the response: choose the first out
of the list.