Update dependencies
Pin the dependency on the lastest procket to fix a build issue on OS X:

https://github.com/msantos/gen_icmp/issues/11

Also update to the latest version of pkt.

Since gen_icmp is still built using rebar2, if procket is pulled in,
it will also be built using rebar2. The plan is to move gen_icmp to
rebar3 after pkt has been converted to use rebar3.

Thanks @kradalby!